(Test): Refactor explainability tests to separate test file [pv622]
