+++ packages/metrics/tests/test_model_intgr.py
@@ -7,7 +7,6 @@ from tests.metric_mocks.mock_model_finite_diff_grad import (
 )
 from metrics.metrics import calculate_metrics
 import numpy as np
-import pytest
 from tests.server_factory import (
     server_factory,
     HOST,
@@ -15,36 +14,6 @@ from tests.server_factory import (
 )
 
 server_factory = server_factory  # To suppress lint errors
-
-
-def test_explanation_stability_similar_scores_result_in_1(server_factory):
-    metric_name = "explanation_stability_score"
-    with server_factory(metric_name):
-        # Check similar predictions after perturbation have value close to 1
-        info = CalculateRequest(
-            metrics=[metric_name],
-            input_features=[[1, 2]],
-            confidence_scores=[[0.5]],
-            model_url=f"http://{HOST}:{server_configs[metric_name]['port']}/predict-10000-ESS",
-            model_api_key="None"
-        )
-        result = calculate_metrics(info)
-        assert result.metric_values[metric_name] == pytest.approx(1.0)
-
-
-def test_explanation_stability_different_scores_is_not_1(server_factory):
-    metric_name = "explanation_stability_score"
-    with server_factory(metric_name):
-        # Check different predictions after perturbation have value close to 0
-        info = CalculateRequest(
-            metrics=[metric_name],
-            input_features=[[1, 2], [3, -4], [-5, 6], [1000, 984], [0, 60], [-34, 2222]],
-            confidence_scores=[[0.5], [0.6], [0.2], [0.8], [0.9], [0.1]],
-            model_url=f"http://{HOST}:{server_configs[metric_name]['port']}/predict-different-ESS",
-            model_api_key="None"
-        )
-        result = calculate_metrics(info)
-        assert result.metric_values[metric_name] < 1.0
 
 
 def test_finite_diff_gradient(server_factory):
@@ -88,8 +57,3 @@ def test_ood_auroc(server_factory):
 
         assert isinstance(result, float), f"Expected AUROC to be a float, but got {type(result)}"
         assert 0.0 <= result <= 1.0, f"Expected AUROC to be between 0.0 and 1.0, but got {result}"
-
-
-# TODO: Test explanation sparse and fidelity metrics
-def test_explanation_sparsity_ideal_case():
-    pass
